+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New outputs:
   - BoM Labels (bom_labels): [Experimental] To generate a PDF containing
     labels for the BoM items. (#886)
+  - PCB stats (pcb_stats): The KiCad 10 PCB statistics
 - CLI:
   - --fail-on-warnings: to return error when we detected warning (See #828)
 - generic filter: `exclude_kicad_dnp` to exclude components marked as DNP in